JCLOUDS-145: Upgrade Abiquo dependencies to the latest version

+++ abiquo/src/test/java/org/jclouds/abiquo/compute/functions/VirtualMachineToNodeMetadataTest.java
@@ -122,7 +122,7 @@ public class VirtualMachineToNodeMetadataTest {
       assertEquals(node.getLocation().getDescription(), "Mock Location");
       assertEquals(node.getImageId(), "1");
       assertEquals(node.getHardware().getId(), "1");
-      assertEquals(node.getHardware().getRam(), vm.getRam());
+      assertEquals(node.getHardware().getRam(), vm.getRam().intValue());
       assertEquals(node.getHardware().getProcessors().get(0).getCores(), (double) vm.getCpu());
       assertEquals(node.getLoginPort(), vm.getVdrpPort());
       assertEquals(node.getPrivateAddresses().size(), 1);
