+++ abiquo/src/test/java/org/jclouds/abiquo/domain/cloud/VirtualMachineLiveApiTest.java
@@ -103,7 +103,7 @@ public class VirtualMachineLiveApiTest extends BaseAbiquoApiLiveApiTest {
          AsyncTask task = vm.update(true);
 
          assertNull(task);
-         assertEquals(vm.getCpu(), 7);
+         assertEquals(vm.getCpu().intValue(), 7);
       } finally {
          env.virtualDatacenter.setCpuCountHardLimit(originalHard);
          env.virtualDatacenter.setCpuCountSoftLimit(originalSoft);
